implemented auth token, secured apis (#7)

+++ routes/category-routes.ts
@@ -1,3 +1,4 @@
+import { verifyToken } from './../helpers/jwt-helper';
 import { NextFunction, Request, Response } from 'express';
 import Category from '../src/app/schema/category';
 
@@ -10,27 +11,39 @@ export class CategoriesRoute {
             });
         });
 
-        app.route('/api/categories').post((req: Request, res: Response, next: NextFunction) => {
-            Category.create(req.body, (err, category) => {
-                if (err) {
-                    return next(err)
-                }
-                res.json(category)
-            })
+        app.route('/api/categories').post(verifyToken, (req: Request, res: Response, next: NextFunction) => {
+            if (!req.isAdmin) {
+                res.status(401).send({ message: "Unauthorized request" })
+            } else {
+                Category.create(req.body, (err, category) => {
+                    if (err) {
+                        return next(err)
+                    }
+                    res.json(category)
+                })
+            }
         })
 
-        app.route('/api/categories/:id').put((req: Request, res: Response, next: NextFunction) => {
-            Category.findByIdAndUpdate(req.params.id, req.body, (err, category) => {
-                if (err) { return next(err); }
-                res.json(category);
-            });
-        });
+        app.route('/api/categories/:id').put(verifyToken, (req: Request, res: Response, next: NextFunction) => {
+            if (!req.isAdmin) {
+                res.status(401).send({ message: "Unauthorized request" })
+            } else {
+                Category.findByIdAndUpdate(req.params.id, req.body, (err, category) => {
+                    if (err) { return next(err); }
+                    res.json(category);
+                })
+            }
+        })
 
-        app.route('/api/categories/:id').delete((req: Request, res: Response, next: NextFunction) => {
-            Category.findByIdAndRemove(req.params.id, req.body, (err, category) => {
-                if (err) { return next(err); }
-                res.json(category);
-            });
-        });
+        app.route('/api/categories/:id').delete(verifyToken, (req: Request, res: Response, next: NextFunction) => {
+            if (!req.isAdmin) {
+                res.status(401).send({ message: "Unauthorized request" })
+            } else {
+                Category.findByIdAndRemove(req.params.id, req.body, (err, category) => {
+                    if (err) { return next(err); }
+                    res.json(category);
+                });
+            }
+        })
     }
 }
