+++ routes/gross-weight-routes.ts
@@ -1,3 +1,4 @@
+import { verifyToken } from './../helpers/jwt-helper';
 import { Request, Response, NextFunction } from 'express'
 import GrossWeight from '../src/app/schema/gross-weight'
 
@@ -10,27 +11,39 @@ export class GrossWeightsRoute {
             });
         });
 
-        app.route('/api/grossWeights').post((req: Request, res: Response, next: NextFunction) => {
-            GrossWeight.create(req.body, (err, grossWeight) => {
-                if (err) {
-                    return next(err)
-                }
-                res.json(grossWeight)
-            })
+        app.route('/api/grossWeights').post(verifyToken, (req: Request, res: Response, next: NextFunction) => {
+            if (!req.isAdmin) {
+                res.status(401).send({ message: "Unauthorized request" })
+            } else {
+                GrossWeight.create(req.body, (err, grossWeight) => {
+                    if (err) {
+                        return next(err)
+                    }
+                    res.json(grossWeight)
+                })
+            }
         })
 
-        app.route('/api/grossWeights/:id').put((req: Request, res: Response, next: NextFunction) => {
-            GrossWeight.findByIdAndUpdate(req.params.id, req.body, (err, grossWeight) => {
-                if (err) { return next(err); }
-                res.json(grossWeight);
-            });
+        app.route('/api/grossWeights/:id').put(verifyToken, (req: Request, res: Response, next: NextFunction) => {
+            if (!req.isAdmin) {
+                res.status(401).send({ message: "Unauthorized request" })
+            } else {
+                GrossWeight.findByIdAndUpdate(req.params.id, req.body, (err, grossWeight) => {
+                    if (err) { return next(err); }
+                    res.json(grossWeight);
+                });
+            }
         });
 
-        app.route('/api/grossWeights/:id').delete((req: Request, res: Response, next: NextFunction) => {
-            GrossWeight.findByIdAndRemove(req.params.id, req.body, (err, grossWeight) => {
-                if (err) { return next(err); }
-                res.json(grossWeight);
-            });
+        app.route('/api/grossWeights/:id').delete(verifyToken, (req: Request, res: Response, next: NextFunction) => {
+            if (!req.isAdmin) {
+                res.status(401).send({ message: "Unauthorized request" })
+            } else {
+                GrossWeight.findByIdAndRemove(req.params.id, req.body, (err, grossWeight) => {
+                    if (err) { return next(err); }
+                    res.json(grossWeight);
+                });
+            }
         });
     }
 }
